Refactor stats logic into modules
